refactor(mir): carve the returned-member read-alias pins out of drop_plan
Threading the callable key through `seal_checked` / `derive_elaboration` put
`drop_plan.rs` over its no-growth line cap. The cap is not to be widened, so
move a coherent block out instead: the returned-member read-alias
re-admission pins become `drop_plan/returned_member_read_aliases.rs`,
alongside the sibling test modules the file already carves out that way.

Pure move — `super` still resolves to `drop_plan`, so every helper the pins
reach is the same one they reached inline.
